revert: undo import refactor (split to separate PR)
Move insightVizDataNodeKey and insightVizDataCollectionId back to
InsightViz.tsx and revert all import paths. The HMR fix is now in a
separate PR (#52385).
