+++ frontend/src/queries/Query/Query.tsx
@@ -8,8 +8,7 @@ import { MarketingAnalyticsOverview } from 'scenes/web-analytics/tabs/marketing-
 import { ErrorBoundary } from '~/layout/ErrorBoundary'
 import { DataNode } from '~/queries/nodes/DataNode/DataNode'
 import { DataTable } from '~/queries/nodes/DataTable/DataTable'
-import { InsightViz } from '~/queries/nodes/InsightViz/InsightViz'
-import { insightVizDataNodeKey } from '~/queries/nodes/InsightViz/utils'
+import { InsightViz, insightVizDataNodeKey } from '~/queries/nodes/InsightViz/InsightViz'
 import { WebOverview } from '~/queries/nodes/WebOverview/WebOverview'
 import { WebVitals } from '~/queries/nodes/WebVitals/WebVitals'
 import { QueryEditor } from '~/queries/QueryEditor/QueryEditor'
