+++ frontend/src/queries/nodes/DataVisualization/DataVisualization.tsx
@@ -13,7 +13,7 @@ import { insightDataLogic } from 'scenes/insights/insightDataLogic'
 import { HogQLBoldNumber } from 'scenes/insights/views/BoldNumber/BoldNumber'
 import { urls } from 'scenes/urls'
 
-import { insightVizDataCollectionId, insightVizDataNodeKey } from '~/queries/nodes/InsightViz/utils'
+import { insightVizDataCollectionId, insightVizDataNodeKey } from '~/queries/nodes/InsightViz/InsightViz'
 import {
     AnyResponseType,
     DataVisualizationNode,
