+++ frontend/src/queries/nodes/InsightViz/InsightViz.tsx
@@ -9,6 +9,7 @@ import { useAttachedLogic } from 'lib/logic/scenes/useAttachedLogic'
 import { insightDataLogic } from 'scenes/insights/insightDataLogic'
 import { insightLogic } from 'scenes/insights/insightLogic'
 import { insightVizDataLogic } from 'scenes/insights/insightVizDataLogic'
+import { keyForInsightLogicProps } from 'scenes/insights/sharedUtils'
 
 import { ErrorBoundary } from '~/layout/ErrorBoundary'
 import { AnyResponseType, DashboardFilter, HogQLVariable, InsightVizNode } from '~/queries/schema/schema-general'
@@ -19,7 +20,16 @@ import { InsightLogicProps } from '~/types'
 import { DataNodeLogicProps, dataNodeLogic } from '../DataNode/dataNodeLogic'
 import { EditorFilters } from './EditorFilters'
 import { InsightVizDisplay } from './InsightVizDisplay'
-import { getCachedResults, insightVizDataCollectionId, insightVizDataNodeKey } from './utils'
+import { getCachedResults } from './utils'
+
+/** The key for the dataNodeLogic mounted by an InsightViz for insight of insightProps */
+export const insightVizDataNodeKey = (insightProps: InsightLogicProps<any>): string => {
+    return `InsightViz.${keyForInsightLogicProps('new')(insightProps)}`
+}
+
+export const insightVizDataCollectionId = (props: InsightLogicProps<any> | undefined, fallback: string): string => {
+    return props?.dataNodeCollectionId ?? props?.dashboardId?.toString() ?? props?.dashboardItemId ?? fallback
+}
 
 type InsightVizProps = {
     uniqueKey?: string | number
