+++ frontend/src/scenes/insights/views/Funnels/FunnelPropertyCorrelationTable.stories.tsx
@@ -8,7 +8,7 @@ import { insightLogic } from 'scenes/insights/insightLogic'
 import { mswDecorator } from '~/mocks/browser'
 import funnelCorrelation from '~/mocks/fixtures/api/projects/team_id/insights/funnelCorrelation.json'
 import { DataNodeLogicProps, dataNodeLogic } from '~/queries/nodes/DataNode/dataNodeLogic'
-import { insightVizDataNodeKey } from '~/queries/nodes/InsightViz/utils'
+import { insightVizDataNodeKey } from '~/queries/nodes/InsightViz/InsightViz'
 import { getCachedResults } from '~/queries/nodes/InsightViz/utils'
 import { InsightLogicProps } from '~/types'
