+++ frontend/src/scenes/notebooks/Nodes/NotebookNodeIssues.tsx
@@ -5,7 +5,7 @@ import { IconX } from '@posthog/icons'
 
 import { useOnMountEffect } from 'lib/hooks/useOnMountEffect'
 
-import { insightVizDataNodeKey } from '~/queries/nodes/InsightViz/utils'
+import { insightVizDataNodeKey } from '~/queries/nodes/InsightViz/InsightViz'
 import { Query } from '~/queries/Query/Query'
 import { InsightLogicProps } from '~/types'
