+++ products/revenue_analytics/frontend/nodes/modals/MRRBreakdownModal.stories.tsx
@@ -6,7 +6,7 @@ import { insightLogic } from 'scenes/insights/insightLogic'
 import { trendsDataLogic } from 'scenes/trends/trendsDataLogic'
 
 import { DataNodeLogicProps, dataNodeLogic } from '~/queries/nodes/DataNode/dataNodeLogic'
-import { insightVizDataNodeKey } from '~/queries/nodes/InsightViz/utils'
+import { insightVizDataNodeKey } from '~/queries/nodes/InsightViz/InsightViz'
 import { InsightLogicProps } from '~/types'
 
 import RevenueAnalyticsMRRQueryMock from '../../__mocks__/RevenueAnalyticsMRRQuery.json'
